add a script to enable the CCNT register on our MPCore box
The problem is that the CCNT register is not incremented unless
specifically enabled on our particlar hardware. Just enabling the
register during boot is insufficient since Linux's performance counter
subsystem claims the hardware and messes with the control flags.

We could either take over the hardware from Linux's performance
counter framework, or trick Linux in turning on the cycle counter for
us. The former approach require caryying ugly kernel patches, the
latter is this simple script...